Reading transfer file bug fixed
Prior to this commit, the second 'send' event would fail because it
attempted to read the TransferAggregate object from the file without
first base64decoding the data. This is now fixed, so that transfers work
correctly from participant 1 onwards.
Also minor fix to the printing of the amount of the fidelity bond to
fund.