Corrects algorithm for detection coin initialized.
Code now checks that precisely the correct set of partial
signatures has been transferred to us, before declaring that the
pathcoin is 'active'. A side effect is that it now correctly only prints
this message out once, instead of spuriously multiple times.
The relevant function in PathCoinParticipantState has a very long
docstring explaining which particular set of partial sigs are needed.