Bugfix: nonce sign flipped status persisted.
Prior to this commit, the adaptor verification on transfer
would sometimes fail, because the reloaded state from disk
was not remembering that the individual nonces signs were flipped,
and so they were being erroneously re-flipped back to their starting
values.
After this commit, we include "has our nonce sign been flipped" as part
of the state that is serialized, avoiding the above error.