Broadcast transactions automatically
For methods spend, reclaim and penalty, we broadcast the transactions
directly over the rpc interface instead of printing them and requiring
the user to manually broadcast them.
Additionally, for localhost testing the user does not need to specify a
list of ports (base port number is in config file).
Also fixed a bug in reclaim, it now pays a fee.